question-generator/apps/api/pyproject.toml
[project]
name = "question-generator-api"
version = "0.1.0"
description = "ACT/SAT question generator API"
authors = [
    {name = "Ethan Rodgers-Gates",email = "erodgersgates@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <4.0"
dependencies = [
    "pymupdf (>=1.26.7,<2.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pdfminer-six (>=20251230,<20251231)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
fix = true

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E203", "E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_configs = true

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.5.1"
alembic = "^1.17.2"
